Add implementation to the Controller

+++ src/main/java/com/peters/coding/task/controller/CaseResourceController.java
@@ -18,37 +18,31 @@ public class CaseResourceController {
 
     @GetMapping("/cases/status/{status}")
     List<Case> getCasesWithStatus(@PathVariable Case.Status status) {
-        // TODO: Query for cases with a specified status.
        return caseService.getCasesWithStatus(status);
     }
 
     @GetMapping("/cases/user/{userId}")
-    List<Case> getOpenCases(@PathVariable Integer userId) {
-        // TODO: Query for cases with a userId;
+    List<Case> getOpenCases(@PathVariable Long userId) {
         return caseService.getOpenCasesByUserId(userId);
     }
 
     @GetMapping("/cases/user/{userId}/status/{status}")
-    List<Case> getOpenCases(@PathVariable Integer userId, @PathVariable Case.Status status) {
-        // TODO: Query for cases with a userId and status
+    List<Case> getOpenCases(@PathVariable Long userId, @PathVariable Case.Status status) {
         return caseService.getOpenCases(userId, status);
     }
 
     @GetMapping("/case/{caseId}")
-    Case getCase(@PathVariable Integer caseId) {
-        // TODO: Get a case by ID
+    Case getCase(@PathVariable Long caseId) {
         return caseService.getCaseById(caseId);
     }
 
     @PostMapping(value = "/case/create", consumes = "application/json")
     Case createCase(@RequestBody Case toCreate) {
-        // TODO: Create a case and any associated records (user, notes)
         return caseService.createCase(toCreate);
     }
 
     @PostMapping(value = "/case/{caseId}/addNote", consumes = "text/plain")
-    Note addNote(@PathVariable Integer caseId, @RequestBody String detail) {
-        // TODO: Create a note for the specified case
+    Note addNote(@PathVariable Long caseId, @RequestBody String detail) {
         return caseService.addNoteToCase(caseId, detail);
     }
 }
